Adds snapshot and story org
